# api/gen/http/openapi.yaml
swagger: "2.0"
info:
    title: Domilike Service
    version: ""
host: localhost:18080
basePath: /api
consumes:
    - application/json
    - application/xml
    - application/gob
produces:
    - application/json
    - application/xml
    - application/gob
paths:
...
    /coins:
        get:
            tags:
                - player-information
            summary: get_coins player-information
            operationId: player-information#get_coins
            parameters:
                - name: Authorization
                  in: header
                  description: JWT token used to perform authorization
                  required: true
                  type: string
            responses:
                "200":
                    description: OK response.
                    schema:
                        type: integer
                        format: int64
            schemes:
                - http
            security:
                - jwt_header_Authorization: []
    /decks:
        get:
            tags:
                - player-information
            summary: get_decks player-information
            operationId: player-information#get_decks
            parameters:
                - name: Authorization
                  in: header
                  description: JWT token used to perform authorization
                  required: true
                  type: string
            responses:
                "200":
                    description: OK response.
                    schema:
                        type: integer
                        format: int64
            schemes:
                - http
            security:
                - jwt_header_Authorization: []
    /discards:
        get:
            tags:
                - player-information
            summary: get_discards player-information
            operationId: player-information#get_discards
            parameters:
                - name: Authorization
                  in: header
                  description: JWT token used to perform authorization
                  required: true
                  type: string
            responses:
                "200":
                    description: OK response.
                    schema:
                        type: integer
                        format: int64
            schemes:
                - http
            security:
                - jwt_header_Authorization: []
...
    /hands:
        get:
            tags:
                - player-information
            summary: get_hands player-information
            operationId: player-information#get_hands
            parameters:
                - name: Authorization
                  in: header
                  description: JWT token used to perform authorization
                  required: true
                  type: string
            responses:
                "200":
                    description: OK response.
                    schema:
                        type: integer
                        format: int64
            schemes:
                - http
            security:
                - jwt_header_Authorization: []
    /play-areas:
        get:
            tags:
                - player-information
            summary: get_play_area player-information
            operationId: player-information#get_play_area
            parameters:
                - name: Authorization
                  in: header
                  description: JWT token used to perform authorization
                  required: true
                  type: string
            responses:
                "200":
                    description: OK response.
                    schema:
                        type: integer
                        format: int64
            schemes:
                - http
            security:
                - jwt_header_Authorization: []
    /player-informations:
        post:
            tags:
                - player-information
            summary: create player-information
            operationId: player-information#create
            parameters:
                - name: Authorization
                  in: header
                  description: JWT token used to perform authorization
                  required: true
                  type: string
            responses:
                "200":
                    description: OK response.
                    schema:
                        type: string
            schemes:
                - http
            security:
                - jwt_header_Authorization: []
        delete:
            tags:
                - player-information
            summary: delete player-information
            operationId: player-information#delete
            parameters:
                - name: Authorization
                  in: header
                  description: JWT token used to perform authorization
                  required: true
                  type: string
            responses:
                "204":
                    description: No Content response.
            schemes:
                - http
            security:
                - jwt_header_Authorization: []
    /victory-points:
        get:
            tags:
                - player-information
            summary: get_victory_points player-information
            operationId: player-information#get_victory_points
            parameters:
                - name: Authorization
                  in: header
                  description: JWT token used to perform authorization
                  required: true
                  type: string
            responses:
                "200":
                    description: OK response.
                    schema:
                        type: integer
                        format: int64
            schemes:
                - http
            security:
                - jwt_header_Authorization: []
securityDefinitions:
    jwt_header_Authorization:
        type: apiKey
        description: JWT based authentication
        name: Authorization
        in: header